+++ execution/function/quantile.go
@@ -13,6 +13,8 @@ package function
 import (
 	"math"
 	"sort"
+
+	"github.com/prometheus/prometheus/model/histogram"
 )
 
 type le struct {
@@ -162,4 +164,81 @@ func ensureMonotonic(buckets buckets) {
 			buckets[i].count = max
 		}
 	}
+}
+
+// histogramQuantile calculates the quantile 'q' based on the given histogram.
+//
+// The quantile value is interpolated assuming a linear distribution within a
+// bucket.
+// TODO(beorn7): Find an interpolation method that is a better fit for
+// exponential buckets (and think about configurable interpolation).
+//
+// A natural lower bound of 0 is assumed if the histogram has only positive
+// buckets. Likewise, a natural upper bound of 0 is assumed if the histogram has
+// only negative buckets.
+// TODO(beorn7): Come to terms if we want that.
+//
+// There are a number of special cases (once we have a way to report errors
+// happening during evaluations of AST functions, we should report those
+// explicitly):
+//
+// If the histogram has 0 observations, NaN is returned.
+//
+// If q<0, -Inf is returned.
+//
+// If q>1, +Inf is returned.
+//
+// If q is NaN, NaN is returned.
+func histogramQuantile(q float64, h *histogram.FloatHistogram) float64 {
+	if q < 0 {
+		return math.Inf(-1)
+	}
+	if q > 1 {
+		return math.Inf(+1)
+	}
+
+	if h.Count == 0 || math.IsNaN(q) {
+		return math.NaN()
+	}
+
+	var (
+		bucket histogram.Bucket[float64]
+		count  float64
+		it     = h.AllBucketIterator()
+		rank   = q * h.Count
+	)
+	for it.Next() {
+		bucket = it.At()
+		count += bucket.Count
+		if count >= rank {
+			break
+		}
+	}
+	if bucket.Lower < 0 && bucket.Upper > 0 {
+		if len(h.NegativeBuckets) == 0 && len(h.PositiveBuckets) > 0 {
+			// The result is in the zero bucket and the histogram has only
+			// positive buckets. So we consider 0 to be the lower bound.
+			bucket.Lower = 0
+		} else if len(h.PositiveBuckets) == 0 && len(h.NegativeBuckets) > 0 {
+			// The result is in the zero bucket and the histogram has only
+			// negative buckets. So we consider 0 to be the upper bound.
+			bucket.Upper = 0
+		}
+	}
+	// Due to numerical inaccuracies, we could end up with a higher count
+	// than h.Count. Thus, make sure count is never higher than h.Count.
+	if count > h.Count {
+		count = h.Count
+	}
+	// We could have hit the highest bucket without even reaching the rank
+	// (this should only happen if the histogram contains observations of
+	// the value NaN), in which case we simply return the upper limit of the
+	// highest explicit bucket.
+	if count < rank {
+		return bucket.Upper
+	}
+
+	rank -= count - bucket.Count
+	// TODO(codesome): Use a better estimation than linear.
+	return bucket.Lower + (bucket.Upper-bucket.Lower)*(rank/bucket.Count)
 }
